Move metadata.py file
